Tests controller files moved to controller folder
