+++ tests/Feature/Controller/DniNieRuleTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Controller;
 
 use App\Rules\DniNieRule;
 use Illuminate\Support\Facades\Validator;
