+++ tests/Feature/Controller/TagControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Controller;
 
 use App\Models\Tag;
 use App\Models\User;
