+++ tests/Feature/Controller/TagListControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature;
+namespace Tests\Feature\Controller;
 
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
